+++ CHANGELOG.md
@@ -6,6 +6,10 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+...
+
+## 7.8.0 (2022-03-10)
+
 * `<cc-env-var-*>`: fix links to the doc
 * `<cc-pricing-product>`: add details for Heptapod runners